Moved unit tests to a separate folder
