+++ src/__tests__/FilterDistrict.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import * as enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import configureMockStore from 'redux-mock-store';
-import FilterDistrict from './FilterDistrict';
+import FilterDistrict from '../Component/Dashboard/FilterDistrict';
 
 let wrapper = null;
