+++ src/__tests__/ModalSession.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import ModalSession from './ModalSession';
+import ModalSession from '../Component/Dashboard/ModalSession';
 
 let wrapper = null;
