+++ src/__tests__/Navigation.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { configure, shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Navigation from './Navigation';
+import Navigation from '../Rooting/Navigation';
 
 let wrapper = null;
