+++ src/__tests__/index.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Setting from './index';
+import Setting from '../Component/Setting/index';
 
 let wrapper = null;
